refactor: streamline type imports and fix type casting issue
- Fixed unnecessary type assertion on `rawBalance` in `use-get-mmc-contract-balance.ts`.
- Consolidated imports from "viem" in `types.ts` for improved readability.
These changes enhance code clarity and reduce potential errors.

+++ src/hooks/protocol/use-get-mmc-contract-balance.ts
@@ -21,7 +21,7 @@ export function useGetMmcContractBalance(address?: Address) {
         args: [address],
       }) as bigint;
 
-      const formattedBalance = parseFloat(formatUnits(rawBalance as bigint, 18));
+      const formattedBalance = parseFloat(formatUnits(rawBalance, 18));
       setBalance(isNaN(formattedBalance) ? 0 : formattedBalance);
     } catch (error) {
       console.error('Error fetching balance:', error);
